tag source: filter by keyword_patterns
This filters the tags by the current `keyword_patterns`.

All tags are read and cached from the tags file, but then the source
will only return the matching/relevant ones.
This will include e.g. `deoplete#foo` in a `ft=vim` file, but not with
`ft=html`.

It also changes the regexp for
`deoplete.sources.base.get_complete_position` to use a non-capturing
group.