Ensure that b:changedtick is unchanged with delayed completion
This adds a short-circuit for the b:changedtick check, but also uses the
tick from when the timer was started.

This is important with manual triggering of completions using e.g.
`<C-p>`, and can be seen best with long timeouts (i.e. `:let
g:deoplete#auto_complete_delay = 1000`): opening of the completion popup
triggers b:changedtick and this prevents deoplete from stomping over the
existing completions.

I have also moved s:completion_delayed down to make it easier to follow
the code.

Apart from the issue this fixes I think that having a short-circuit for
the common case of checking b:changedtick is good to have - skipping the
building of the context all the time.