Improve error message when loading fails
This will result in the following message in case you have a SyntaxError
in the Python host:

> [deoplete] deoplete failed to load: Failed to load python3 host. You can try to see what happened by starting nvim with $NVIM_PYTHON_LOG_FILE set and opening the generated log file. Also, the host stderr is available in messages.. Try the :UpdateRemotePlugins command and restart Neovim.

I think it is useful to have `v:exception` in there, and to only use a
single line.  Before this, you would only see the 2nd error/line, and
there was no context about the error.